+++ core/schema/user-schema.js
@@ -6,15 +6,24 @@ const userSchema = new mongoose.Schema({
     required: true
   },
   address: {
-    flat_details: {
+    addressLine1: {
       type: String
     },
-    area: {
+    addressLine2: {
       type: String
     },
-    landmark: {
+    city: {
       type: String
-    }
+    },
+    state: {
+      type: String
+    },
+    country: {
+      type: String
+    },
+    postalCode: {
+      type: String
+    },
   },
   age: {
     type: Number,
