+++ index.js
@@ -2,17 +2,15 @@ require('dotenv').config();
 const express = require('express');
 const cors = require('cors');
 const app = express();
-const mongoose = require('mongoose')
+const mongoose = require('mongoose');
 const swaggerUi = require('swagger-ui-express');
 const swaggerSpecs = require('./swagger');
-const registerRoute = require('./routes/registerRoute')
-const loginRoute = require('./routes/loginRoute');
+const authRoute = require('./routes/authRoute');
 
 app.use(cors());
 app.use(express.json());
 app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(swaggerSpecs));
-app.use("/register", registerRoute);
-app.use("/login", loginRoute);
+app.use("/auth", authRoute);
 
 mongoose.connect(process.env.MONGODB_URI)
   .then(async function () {
