+++ middlewares/authenticateToken.js
@@ -1,0 +1,20 @@
+const jwt = require('jsonwebtoken');
+const secretKey = require('../core/constant/jwtKeys');
+
+const authenticateToken = (req, res, next) => {
+    let token = req.headers.authorization?.split(' ')[1];
+    if (!token) {
+        return res.status(401).json({ error: 'Unauthorized: Token missing' });
+    }
+
+    jwt.verify(token, secretKey, (err, user) => {
+
+        if (err) {
+            return res.status(403).json({ error: 'Forbidden: Invalid token' });
+        }
+        req.user = user;
+        next();
+    });
+};
+
+module.exports = { authenticateToken }
